+++ packages/apps/stacc-website/src/Routes.js
@@ -9,17 +9,15 @@ import FourOFour from "./pages/404"
 import styled from "styled-components"
 
 const Routes = epitath(function*({ change, inverted }) {
-  const { result: clientCaseStudyPages } = yield (
-    <SanityList type={"client"} filter={`hide != true`} />
-  )
   const { result: pages } = yield (
     <SanityList
-      type={`page" || _type == "servicePage" || _type == "productPage" || _type == "teamPage" || _type == "clientsPage" || _type == "campaignPage`}
+      filter={`_type match "*page"`}
       pick={"blocks[]{employee->,...},..."}
+      id={"pages"}
     />
   )
 
-  if (!pages || !clientCaseStudyPages) {
+  if (!pages) {
     return (
       <LoadingContainer>
         <Loading />
@@ -36,51 +34,30 @@ const Routes = epitath(function*({ change, inverted }) {
       />
       <Route path="/" exact render={() => <Home change={change} />} />
       <Route path="/404" exact component={FourOFour} />
-      <Route path={"/clients"}>
-        <Switch>
-          <Route path={"/clients"} exact component={Clients} />
-          {clientCaseStudyPages.map(client => {
-            return client.caseStudies
-              ? client.caseStudies.map(caseStudy => {
-                  if (!caseStudy.path) {
-                    return null
-                  }
-                  return (
-                    <Route
-                      key={caseStudy._key}
-                      exact
-                      path={caseStudy.path.current}
-                      render={() => <Case caseStudy={caseStudy} />}
-                    />
-                  )
-                })
-              : null
-          })}
+      <Route path={"/clients"} exact component={Clients} />
+      {pages.map(page => {
+        const exact = !!(!page.subpages || page.subpages.length <= 0)
+        return (
           <Route
-            render={() => {
-              if (!inverted) {
-                change({ inverted: true })
+            key={page._id}
+            path={`${page.path.current}`}
+            render={({ match, location }) => {
+              if (page._type === "caseStudyPage") {
+                return <Case caseStudy={page} />
               }
-              return <FourOFour />
+              return (
+                <Page
+                  page={page}
+                  match={match}
+                  location={location}
+                  change={change}
+                />
+              )
             }}
+            exact={exact}
           />
-        </Switch>
-      </Route>
-      {pages.map(page => (
-        <Route
-          key={page._id}
-          path={`${page.path.current}`}
-          render={({ match, location }) => (
-            <Page
-              page={page}
-              match={match}
-              location={location}
-              change={change}
-            />
-          )}
-        />
-      ))}
-      }}
+        )
+      })}
       <Route
         render={() => {
           if (!inverted) {