+++ packages/themes/Fundu/src/Theme/Reset.js
@@ -1,0 +1,11 @@
+import { css } from "styled-components"
+import { theming } from "@staccx/base"
+
+export default theming.createGlobal({
+  baseReset: theming.cssResets,
+  other: css`
+    input:-webkit-autofill {
+      background-color: ${theming.color.bg} !important;
+    }
+  `
+})